# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildPro Construction Business Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/history:
    get:
      operationId: getClientHistory
      tags: [clients]
      summary: Get customer job history and totals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientHistory"

  # ──────────────────────────────────────
  # JOBS
  # ──────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: [string, "null"]
        phone:
          type: [string, "null"]
        address:
          type: [string, "null"]
        city:
          type: [string, "null"]
        state:
          type: [string, "null"]
        zipCode:
          type: [string, "null"]
        notes:
          type: [string, "null"]
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: [string, "null"]
        phone:
          type: [string, "null"]
        address:
          type: [string, "null"]
        city:
          type: [string, "null"]
        state:
          type: [string, "null"]
        zipCode:
          type: [string, "null"]
        notes:
          type: [string, "null"]
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: [string, "null"]
        phone:
          type: [string, "null"]
        address:
          type: [string, "null"]
        city:
          type: [string, "null"]
        state:
          type: [string, "null"]
        zipCode:
          type: [string, "null"]
        notes:
          type: [string, "null"]
    ClientHistory:
      type: object
      required: [clientId, totalJobs, completedJobs, totalRevenue, jobs]
      properties:
        clientId:
          type: integer
        totalJobs:
          type: integer
        completedJobs:
          type: integer
        totalRevenue:
          type: number
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
    Job:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: [integer, "null"]
        clientName:
          type: [string, "null"]
        status:
          type: string
        jobType:
          type: [string, "null"]
        priority:
          type: string
        description:
          type: [string, "null"]
        address:
          type: [string, "null"]
        city:
          type: [string, "null"]
        state:
          type: [string, "null"]
        zipCode:
          type: [string, "null"]
        startDate:
          type: [string, "null"]
        endDate:
          type: [string, "null"]
        estimatedValue:
          type: [number, "null"]
        actualCost:
          type: [number, "null"]
        notes:
          type: [string, "null"]
        createdAt:
          type: string
...